#27 Created Sharer utils
