+++ capacitor/plugins-library/src/models/KrikooMimeType.ts
